+++ client/src/assets/main.css
@@ -147,7 +147,7 @@
 
 .send{
   margin-left:85vw;
-  margin-top: -34px;
+  margin-top: -54px;
 }
 
 .codestyle{
